requirements.txt
aerich==0.7.1
aiogram==3.0.0b6
toml==0.10.2
tortoise-orm[asyncpg]==0.19.3
click==8.1.3
typing==3.7.4.3
